created useUserTransactionsUpdates to handle all transactions updates. possible conflicting/duplicate behavior with useFinalizeUserTransactions, will most likely look to combine them.
