+++ packages/augur-comps/src/index.tsx
@@ -58,6 +58,7 @@ import {
   keyedObjToKeyArray,
   arrayToKeyedObject,
   arrayToKeyedObjectByProp,
+  useUserTransactionsUpdates,
 } from './stores/utils';
 import * as _ApprovalHooks from './stores/use-approval-callback';
 import * as _GraphClient from './apollo/client';
@@ -83,6 +84,7 @@ export const Stores = {
     useFinalizeUserTransactions,
     useScrollToTopOnMount,
     useLocalStorage,
+    useUserTransactionsUpdates,
     ..._ApprovalHooks,
   },
   Utils: {
@@ -158,6 +160,7 @@ export {
   useCanEnterCashPosition,
   useUserBalances,
   useFinalizeUserTransactions,
+  useUserTransactionsUpdates,
   useScrollToTopOnMount,
   useGraphDataStore,
   GraphDataStore,
