+++ packages/augur-comps/src/stores/user-hooks.ts
@@ -4,7 +4,6 @@ import { windowRef } from '../utils/window-ref';
 import { USER_ACTIONS, USER_KEYS, DEFAULT_USER_STATE } from './constants';
 import { UserBalances, TransactionDetails } from '../utils/types';
 import { augurSdkLite } from '../utils/augurlitesdk';
-import { TX_STATUS } from '../utils/constants';
 
 const {
   ADD_TRANSACTION,
