+++ packages/augur-comps/src/stores/utils.ts
@@ -304,18 +304,34 @@ export function useScrollToTopOnMount(...optionsTriggers) {
   }, [...optionsTriggers]);
 }
 
-export function useUpdateTXStatus(txResponse) {
-  const {
-    transactions: tx,
-    actions: { updateTransaction },
-  } = useUserStore();
-  if (txResponse.confirmations > 0) {
-    const updateTxState = tx.find(
-      (tx) => tx.hash === txResponse.transactionHash
-    );
-    if (updateTxState) {
-      updateTxState.status = TX_STATUS.CONFIRMED;
-      updateTransaction(txResponse.transactionHash, updateTxState);
+export function useUserTransactionsUpdates() {
+  const { blocknumber } = useGraphDataStore();
+  const { loginAccount, transactions, actions: { updateTransaction } } = useUserStore();
+  useEffect(() => {
+    if (blocknumber && transactions) {
+      transactions
+        .filter((tx) => tx?.status === TX_STATUS.PENDING)
+        .forEach((tx) => {
+          const isTransactionMined = async (transactionHash, provider) => {
+            const txReceipt = await provider.getTransactionReceipt(
+              transactionHash
+            );
+            if (txReceipt && txReceipt.blockNumber) {
+              return txReceipt;
+            }
+          };
+          isTransactionMined(tx.hash, loginAccount.library).then((response) => {
+            if (response?.confirmations > 0) {
+              const updateTxState = transactions.find(
+                (tx) => tx.hash === response.transactionHash
+              );
+              if (updateTxState) {
+                updateTxState.status = TX_STATUS.CONFIRMED;
+                updateTransaction(response.transactionHash, updateTxState);
+              }
+            }
+          });
+        });
     }
-  }
+  }, [transactions, blocknumber]);
 }
